+++ attributes/default.rb
@@ -2,7 +2,7 @@
 # Cookbook Name:: confluence
 # Attributes:: confluence
 #
-# Copyright 2012-2013
+# Copyright 2013, Brian Flad
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -17,13 +17,98 @@
 # limitations under the License.
 #
 
-default['confluence']['version']        = "4.3.7"
-default['confluence']['arch']           = "x64"
-default['confluence']['url']            = "http://www.atlassian.com/software/confluence/downloads/binary/atlassian-confluence-#{node['confluence']['version']}-#{node['confluence']['arch']}.bin"
-default['confluence']['checksum']       = "03bb665b6abdc96495fca3f12683083d5e4633152f0c5cb4464779dcdd869f71"
-default['confluence']['install_path']   = "/opt/atlassian/confluence"
 default['confluence']['home_path']      = "/var/atlassian/application-data/confluence"
+default['confluence']['install_path']   = "/opt/atlassian/confluence"
+default['confluence']['install_type']   = "installer"
+default['confluence']['url_base']       = "http://www.atlassian.com/software/confluence/downloads/binary/atlassian-confluence"
 default['confluence']['user']           = "confluence"
+default['confluence']['version']        = "4.3.7"
+
+if node['kernel']['machine'] == "x86_64"
+  default['confluence']['arch'] = "x64"
+else
+  default['confluence']['arch'] = "x32"
+end
+
+case node['platform_family']
+when "windows"
+  case node['confluence']['install_type']
+  when "cluster-standalone"
+    default['confluence']['url']      = "#{node['confluence']['url_base']}-#{node['confluence']['version']}-cluster.zip"
+    default['confluence']['checksum'] = case node['confluence']['version']
+    when "4.3.7"; "c9f9213a787d76151ef133d75247707300cb55e93c28961276b8c6c4f702b3bf"
+    end
+  when "cluster-war"
+    default['confluence']['url']      = "#{node['confluence']['url_base']}-#{node['confluence']['version']}-cluster-war.zip"
+    default['confluence']['checksum'] = case node['confluence']['version']
+    when "4.3.7"; "92176b0f398cc6b2c6fc9def18268bb82e85361ae297de1a4380d10d0b55c4a6"
+    end
+  when "installer"
+    default['confluence']['url']      = "#{node['confluence']['url_base']}-#{node['confluence']['version']}-#{node['confluence']['arch']}.exe"
+    default['confluence']['checksum'] = case node['confluence']['version']
+    when "4.3.7"; node['confluence']['arch'] == "x64" ? "5cc58d87f82afae574f053135ee214719c5e46f689affc1bdb819f69c69ed216" : "5316a2a4b35efc0342fee27044ca65f3eff1103ac4f09fb3d11596941dd00b55"
+    end
+  when "standalone"
+    default['confluence']['url']      = "#{node['confluence']['url_base']}-#{node['confluence']['version']}.zip"
+    default['confluence']['checksum'] = case node['confluence']['version']
+    when "4.3.7"; "726053c9ed8f56100bb5b3eb0ca9430e1beab73c3e26488cda1d23a42dd52dd7"
+    end
+  when "war"
+    default['confluence']['url']      = "#{node['confluence']['url_base']}-#{node['confluence']['version']}-war.zip"
+    default['confluence']['checksum'] = case node['confluence']['version']
+    when "4.3.7"; "f97cb7ca185a975769c787248784fb36d65645bc4ed70dbbac4a5b7c7ce4fc59"
+    end
+  end
+else
+  case node['confluence']['install_type']
+  when "cluster-standalone"
+    default['confluence']['url']      = "#{node['confluence']['url_base']}-#{node['confluence']['version']}-cluster.tar.gz"
+    default['confluence']['checksum'] = case node['confluence']['version']
+    when "4.3.7"; "d3ca45a1358166c67bf332bceef68393195a2bf800b4ee4fea18159144dfb4eb"
+    end
+  when "cluster-war"
+    default['confluence']['url']      = "#{node['confluence']['url_base']}-#{node['confluence']['version']}-cluster-war.tar.gz"
+    default['confluence']['checksum'] = case node['confluence']['version']
+    when "4.3.7"; "b644b8a95edd3a79b9406c828852174133d6f18d3f853b724d3724e16d5615de"
+    end
+  when "installer"
+    default['confluence']['url']      = "#{node['confluence']['url_base']}-#{node['confluence']['version']}-#{node['confluence']['arch']}.bin"
+    default['confluence']['checksum'] = case node['confluence']['version']
+    when "4.3.7"
+      node['confluence']['arch'] == "x64" ? "03bb665b6abdc96495fca3f12683083d5e4633152f0c5cb4464779dcdd869f71" : "6612ab99ae0cf3ab240f9d9413a25bfe84b3f729cbb12ee4bee4e11a424513d0"
+    end
+  when "standalone"
+    default['confluence']['url']      = "#{node['confluence']['url_base']}-#{node['confluence']['version']}.tar.gz"
+    default['confluence']['checksum'] = case node['confluence']['version']
+    when "4.3.7"; "18602095e4119cc5498ac14b4b5588e55f707cca73b36345b95c9a9e4c25c34d"
+    end
+  when "war"
+    default['confluence']['url']      = "#{node['confluence']['url_base']}-#{node['confluence']['version']}-war.tar.gz"
+    default['confluence']['checksum'] = case node['confluence']['version']
+    when "4.3.7"; "2f1b3a074f083dc4cd8f38505e4fc75cab1961a8880192512482983f56aaa0d7"
+    end
+  end
+end
+
+default['confluence']['apache2']['access_log']         = ""
+default['confluence']['apache2']['error_log']          = ""
+default['confluence']['apache2']['port']               = 80
+default['confluence']['apache2']['virtual_host_alias'] = node['fqdn']
+default['confluence']['apache2']['virtual_host_name']  = node['hostname']
+
+default['confluence']['apache2']['ssl']['access_log']       = ""
+default['confluence']['apache2']['ssl']['chain_file']       = ""
+default['confluence']['apache2']['ssl']['error_log']        = ""
+default['confluence']['apache2']['ssl']['port']             = 443
+
+case node['platform_family']
+when 'rhel'
+  default['confluence']['apache2']['ssl']['certificate_file'] = "/etc/pki/tls/certs/localhost.crt"
+  default['confluence']['apache2']['ssl']['key_file']         = "/etc/pki/tls/private/localhost.key"
+else
+  default['confluence']['apache2']['ssl']['certificate_file'] = "/etc/ssl/certs/ssl-cert-snakeoil.pem"
+  default['confluence']['apache2']['ssl']['key_file']         = "/etc/ssl/private/ssl-cert-snakeoil.key"
+end
 
 default['confluence']['database']['host']     = "localhost"
 default['confluence']['database']['name']     = "confluence"