+++ files/default/tests/minitest/linux_war_test.rb
@@ -1,8 +1,8 @@
 require File.expand_path('../support/helpers', __FILE__)
 
-describe_recipe "confluence::default" do
+describe_recipe "confluence::linux_war" do
   include Helpers::Confluence
 
-  # check other recipe tests
+  # do work, son.
   
 end